Initial structure
Five daemons are planned for the initial version of the lightswitch:
* lightsd - This daemon controls the actual lights and pushes data to
fadecandy server.
* stated - This daemon interacts with external sources of user input and
coordinates the state of the system, the communication thereof and is
hopefully the only service which will pass control sequences to lightsd
* ambd - This daemon keeps track of ambient state of the default effect,
including blueshift adjustments according to time of day and periodic
light fluctuations just to keep things special.
* switchd - This daemon takes input from the actual switch (knob)
mounted on the wall and turns that into lighting changes.
* webd - This daemon provides a web interface to the system and allows
users to control the system via the web.

We use a classic recursive Makefile with apologies to the late Peter
Miller who sought to bring enlightenment to the heretics.  However, for
the current size of the project and expected level of complete
separation betweeen the binaries involved, a recursive Makefile appears
to be a somewhat reasonable approach.  I anticipate I will regret this
decision later.

This commit also updates the .gitignore to appropriately ignore the
binaries generated by the make sequence.